# docs/swagger.yaml
definitions:
...
  employerprofile.EmployerProfile:
    type: object
...
  notification.NotificationPreferences:
    properties:
      application_updates:
        type: boolean
      company_news:
        type: boolean
      created_at:
        type: string
      daily_job_matches:
        type: boolean
      email_notifications:
        type: boolean
      id:
        type: string
      job_alerts:
        type: boolean
      marketing_emails:
        type: boolean
      push_notifications:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: string
      weekly_digest:
        type: boolean
    type: object
  notification.NotificationPreferencesResponse:
    properties:
      data:
        $ref: '#/definitions/notification.NotificationPreferences'
      message:
        type: string
      success:
        type: boolean
    type: object
  notification.UpdateNotificationPreferencesRequest:
    properties:
      application_updates:
        type: boolean
      company_news:
        type: boolean
      daily_job_matches:
        type: boolean
      email_notifications:
        type: boolean
      job_alerts:
        type: boolean
      marketing_emails:
        type: boolean
      push_notifications:
        type: boolean
      weekly_digest:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/employers/{employerId}/profile:
    delete:
      consumes:
      - application/json
      description: Delete a specific employer profile by ID
      parameters:
      - description: Employer ID
        in: path
        name: employerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile deleted successfully
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Employer Profile
      tags:
      - Employer Profile
    get:
      consumes:
      - application/json
      description: Get a specific employer profile by ID
      parameters:
      - description: Employer ID
        in: path
        name: employerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Profile fetched successfully
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Employer profile not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Employer Profile
      tags:
      - Employer Profile
    put:
      consumes:
      - application/json
      description: Update a specific employer profile by ID
      parameters:
      - description: Employer ID
        in: path
        name: employerId
        required: true
        type: string
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employerprofile.EmployerProfile'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Employer Profile
      tags:
      - Employer Profile
  /api/employers/me/profile:
    get:
      consumes:
      - application/json
      description: Get the current user's employer profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile fetched successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Profile not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get My Employer Profile
      tags:
      - Employer Profile
  /api/employers/profile:
    post:
      consumes:
      - application/json
      description: Create a new employer profile for the current user
      parameters:
      - description: Profile creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employerprofile.EmployerProfile'
      produces:
      - application/json
      responses:
        "201":
          description: Profile created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Employer Profile
      tags:
      - Employer Profile
...
  /api/notifications/email:
    post:
      consumes:
      - application/json
      description: Send an email notification
      parameters:
      - description: Email data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Email sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Send Email
      tags:
      - Notifications
  /api/notifications/preferences:
    get:
      consumes:
      - application/json
      description: Get notification preferences for the current user
      produces:
      - application/json
      responses:
        "200":
          description: Preferences retrieved successfully
          schema:
            $ref: '#/definitions/notification.NotificationPreferencesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Notification Preferences
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: Update notification preferences for the current user
      parameters:
      - description: Notification preferences
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notification.UpdateNotificationPreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Preferences updated successfully
          schema:
            $ref: '#/definitions/notification.NotificationPreferencesResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Notification Preferences
      tags:
      - Notifications
...
  /api/worker/job:
    post:
      consumes:
      - application/json
      description: Enqueue a background job for processing
      parameters:
      - description: Job data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Job enqueued successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Queue is full
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Enqueue Background Job
      tags:
      - Background Jobs
swagger: "2.0"